Remove LinkListStore
