+++ common/components/forms/LinkList.jsx
@@ -9,11 +9,6 @@ import type { Dictionary, KeyValuePair } from "../types/Generics.jsx";
 import Sort from "../utils/sort.js";
 import stringHelper from "../utils/string.js";
 import UniversalDispatcher from "../stores/UniversalDispatcher.js";
-import {
-  generateLinkKey,
-  NewLinkInfo,
-  linkCaptions,
-} from "../stores/LinkListStore.js";
 import FormFieldsStore from "../stores/FormFieldsStore.js";
 import _ from "lodash";
 import { createDictionary } from "../types/Generics.jsx";
@@ -21,6 +16,20 @@ import TextFormField, { TextFormFieldType } from "./fields/TextFormField.jsx";
 import { FormFieldValidator } from "../utils/validation.js";
 import urlHelper from "../utils/url.js";
 import Visibility from "../common/Visibility.jsx";
+import { LinkTypes } from "../constants/LinkConstants";
+
+export const linkCaptions: Dictionary<string> = _.fromPairs([
+  [LinkTypes.CODE_REPOSITORY, "Code Repository (e.g. Github)"],
+  [LinkTypes.MESSAGING, "Communication (e.g. Slack)"],
+  [LinkTypes.PROJECT_MANAGEMENT, "Project Management (e.g. Trello)"],
+  [LinkTypes.FILE_REPOSITORY, "File Repository (e.g. Google Drive)"],
+  [LinkTypes.DESIGN, "Design Files (e.g. Figma)"],
+  [LinkTypes.TWITTER, "Twitter"],
+  [LinkTypes.FACEBOOK, "Facebook"],
+  [LinkTypes.LINKED_IN, "LinkedIn"],
+]);
+
+export type NewLinkInfo = {| tempId: ?string |} & LinkInfo;
 
 type Props = {|
   linkOrdering: $ReadOnlyArray<string>,
