+++ common/components/stores/UniversalDispatcher.js
@@ -5,7 +5,6 @@
 import { Dispatcher } from "flux";
 import type { NavigationActionType } from "./NavigationStore.js";
 import type { OffsetActionType } from "./PageOffsetStore.js";
-import type { LinkListActionType } from "./LinkListStore.js";
 import type { FormFieldsActionType } from "./FormFieldsStore.js";
 import type { FavoritesActionType } from "./FavoritesStore.js";
 import type { ProjectSearchActionType } from "./ProjectSearchStore.js";
@@ -13,7 +12,6 @@ import type { ProjectSearchActionType } from "./ProjectSearchStore.js";
 type UniversalActionType =
   | NavigationActionType
   | OffsetActionType
-  | LinkListActionType
   | FormFieldsActionType
   | FavoritesActionType
   | ProjectSearchActionType;
